t/linux: patch out static virtual interfaces
A few virtual drivers (bonding, dummy) predate netlink (~2003), which
means that the kernel had no way to dynamically create network
interfaces.
The solution was kernel module paramter which statically precreated a
list of these virtual interfaces. The number was generally set to 1 by
default, meaning that loading the module creates one of its interface.

For compatibility with legacy userspaces this is still kept around. We
could set the parameters to zero, but doing that everywhere is a pain.
This just patches the default values to zero.

Change-Id: I605781b80fb8b20a7724e7fdfa5a4f75ca25eea1
Reviewed-on: https://review.monogon.dev/c/monogon/+/1589
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>