m/n/c/cluster: set StorageSecurity in bootstrap path
Currently this is only set properly in the register path, causing the
configuration of the bootstrap node to be inconsistent. This causes
bootup to fail with `Node startup failed: sealed configuration has
invalid node unlock key (wanted 32 bytes, got 0)`.
Fix this by also persisting the chosen storage security option in the
node configuration when bootstrapping.

Change-Id: I93bf75d412c9aa7c09b5a739ce65b6873d947fd5
Reviewed-on: https://review.monogon.dev/c/monogon/+/1815
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>