m/n/c/curator: clean up stale leader election after reboot
This is useful if we reboot a leader and it comes back while its' old
stale leader election key/value is present.

Without this, other nodes would continue to connect to the newly
rebooted leader even though it is not a leader anymore, confusing
themselves and cluster operators in the process.

Change-Id: I306e7040550084ef39ab20c3c289a3137145a2d9
Reviewed-on: https://review.monogon.dev/c/monogon/+/1845
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI