m/node: enlarge K8s networks
For bigger clusters, the current 10.0.0.0/16 subnet is far too small.
Switch to 10.192.0.0/11 which should be out of the way of most of our
test infra and is large enough for 8192 nodes with 253 pods which is
big enough for the time being. Also migrate the service network
to 10.224.0.0/16 and make it much bigger. It does not need to be in the
pod CIDR, so move it out of there.
But for large clusters this will continue to be a problem until we have
a better allocation algorithm or switch to IPv6 with 464xlat (which
is not supported on Linux currently however).

Change-Id: Ib3a019fffacec2172721f04c01133b44bffba73b
Reviewed-on: https://review.monogon.dev/c/monogon/+/1848
Reviewed-by: Leopold Schabel <leo@monogon.tech>
Tested-by: Jenkins CI