m/n/k8s: fix start after unclean shutdown
Both the kvmdevice as well as the CSI runnables listen on Unix sockets.
These are normally removed on close (this is actually the default for
sockets opened wiht ListenUnix, thus drop setting this), but when an
unclean shutdown occurs they persist. Since one cannot listen on an
already-existing socket, opportunistically remove them before listening.

Change-Id: I11d986a2816fde3d7ffef0817ae3bbf39bba4faf
Reviewed-on: https://review.monogon.dev/c/monogon/+/1867
Tested-by: Jenkins CI
Reviewed-by: Leopold Schabel <leo@monogon.tech>