m/n/core/curator: fix clusternet sync issues
Clusternet sync was broken whenever a node just started a curator
watcher, as the curator's codepath to serve the backlog wasn't copying
over clusternet data.

This shouldn't have happened, especially as we implemented a unified
function to convert node data into node update data, we just forgot to
use it during the initial backlog generation code on the curator.

I've spent some time trying to come up with a testcase that would
automatically catch any further bug of this type, but that's not really
doable without having more formalized type casts between all the
different types a node can be encoded in (curator in-memory, curator
proto state, api node object). But we do still update one of the curator
tests to catch this particular regression.

Change-Id: I203d9a41b735db63d076c7e68a9fc6fe2f795ab4
Reviewed-on: https://review.monogon.dev/c/monogon/+/1912
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI