metropolis/test: use localregistry
This removes everything but the preseed test image from the preseed
image pool, instead opting to serve all test image via localregistry.

The registry API is served from a dedicated IP inside the virtual
network and forwarded to an ephemeral listener on the host. The relevant
infrastructure is added to the launch package.

As it is required to add configuration to containerd for this registry
anyways as it does not and should not have TLS we take that opportunity
to give it a descriptive name (test.monogon.internal).

Visibilities of images are also adjusted as they are now referenced much
closer to their point of use.

Against main this saves 51MiB in bundle size (289MiB -> 238MiB).

Change-Id: I31f732eb8c4ccec486204f35e3635b588fd9c85b
Reviewed-on: https://review.monogon.dev/c/monogon/+/1927
Tested-by: Jenkins CI
Reviewed-by: Leopold Schabel <leo@monogon.tech>