metropolis/node: dont allocate node ports in constructor
All good things are three, and this should be the final installment in
the saga of getting ports reserved. The Kubernetes codebase always
starts a recovery after initialization of a new port allocator, which
currently makes it very unhappy as we already allocated the system
ports, but we are trying to do the same in the recovery by pretending to
be a service. Anyway, this removes the initial part of the patch and
only uses the recovery way of reserving the ports. This still creates an
annoying message sometimes, but I can't find the code path which creates
them.

Change-Id: Ib7d9ec5d00cbde7371d876c31c63b5312024a187
Reviewed-on: https://review.monogon.dev/c/monogon/+/2027
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI