metropolis/hostsfile: update our own node's roles correctly
This properly serializes the state of our own node's roles into startup
data (and the hosts file) if it is assigned during node runtime.

Without this, we could end up serializing out of date node role data and
thus would not startup up with a cluster directory (and persisted node
roles) which would allow for proper control plane startup.

Change-Id: I45da5d73cc0eea0a7c32308bbecf512bb9699d55
Reviewed-on: https://review.monogon.dev/c/monogon/+/2068
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>