m/n/c/consensus: fix flaky TestEtcdMetrics
The metrics endpoint comes up before the bootstrap process is complete.
The test context then gets cancelled as the test has succeeded,
but the consensus runnable is not done bootstrapping, causing it to
return an error and TestHarness to
fail the test.

Fix this by waiting for the etcd state to be populated, indicating that
the bootstrap machinery has done its job.

Fixes https://github.com/monogon-dev/monogon/issues/258

Change-Id: Ied270191aaebb226822a1c9d7f8c6312bd0da1ed
Reviewed-on: https://review.monogon.dev/c/monogon/+/2043
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>