metropolis/curator: report leader based on ledership election status
This makes GetCurrentLeader calls respond to cluster/leader changes
instead of relying on connections getting closed on time when a leader
changes, or when a node gets elected as a leader.

This leads to less churn when cluster leadership status changes (no need
for every RPC client to re-establish connectivity to get the new
leader). It should also be faster to respond to cluster leader changes,
as it doesn't rely on the the RPC client detecting that the node it
connected to has stopped responding / disconnected.

Change-Id: I9de12286530226b3832d2ae07cb7d943ca537d3f
Reviewed-on: https://review.monogon.dev/c/monogon/+/2069
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>