m/n/c/curator/watcher: init
curator/watcher is a new library which simplifies access to the Curator
from client code.

It implements common logic found in a bunch of Metropolis code, where a
node or multiple nodes are watched for updates. The single-node version
of this RPC is not that complex, but the many-node version is and
definitely deserves a client library to remove possible sources of
implementation bugs.

This doesn't yet switch over any of the existing code to use this
library - that will come up in another CR on the stack.

Change-Id: Ia1df6f21473f86cb9af1e962d157350a5cfd7d07
Reviewed-on: https://review.monogon.dev/c/monogon/+/2262
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI