m/n/c/n/hostsfile: do not log individual entries
This gets extremely spammy when spinning up large clusters as it
emits all entries for every new hostsfile into the log.
For a fresh cluster of n nodes this is O(n^2) log lines, which is
far too many. The number of entries is already logged and should be
sufficient for debugging purposes.

Change-Id: I6bf4e489aea52bad3a32a8a0f900e832b6a4f197
Reviewed-on: https://review.monogon.dev/c/monogon/+/2298
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>