version/spec: init
This is the second step in a repo-wide versioning system after the new
build status generator.

It introduces a proto representation of the essential version data which
is attached to artifacts generated from this repository.

The definition encompasses an informal spec on how this data can be
serialized to a SemVer 2.0.0 version string, which will be used in human
readable output and with interoperability with external systems.

Change-Id: I3f78e34af1f063fcf6035a0258a383b5d3f8adb6
Reviewed-on: https://review.monogon.dev/c/monogon/+/2331
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI