m/n/c/network: add quirks infra and i40e quirk
This adds the applyQuirks function which is called during early
initialization of the network stack, before any network interfaces are
enabled. This function applies device and/or driver-specific fixups to
make them work better. For that purpose it examines relevant metadata
(driver in use, firmware version and OpROM version) to decide which
quirks should be applied to which device.

As we do not yet have another way of exposing firmware versions, this
also takes care of logging all non-zero firmware versions.

It also adds a first quirk for i40e which disables firmware-based LLDP
procesing.

Change-Id: I456753880102dfb5b3e94847cf5627a003d70eeb
Reviewed-on: https://review.monogon.dev/c/monogon/+/2344
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>
Tested-by: Jenkins CI