m/node: panic on unknown NMIs
Configure the kernel to panic in case it receives an unknown NMI.

This can be used to induce a crash in a hanging kernel, allowing
crash diagnostics (pstore currently, maybe kdump in the future) to be
used to diagnose the hang.

This can have false positives on some badly-configured client systems,
but is expected to be fine on servers as other server kernels like
ESXi's vmkernel behave the same.

Change-Id: Ie41e7b816be8492636b6fa596c1c299c0a7a0fc2
Reviewed-on: https://review.monogon.dev/c/monogon/+/2754
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>