m/n/c/clusternet: add clusternet node ip to lo
Without this address the kernel will send all traffic through clusternet
will with the external IP as source address, which prevents any access
from the k8s api-server to pods. By adding the address to the lo
interface, the kernel decides that this is the best match for the
generic clusternet route and uses it as source address.

Before: 10.192.1.2 dev clusternet src 233.252.0.1
After: 10.192.1.2 dev clusternet src 10.192.0.1

Change-Id: I1a78de9d4cc043f93398fe5b6cb10dd202bf75fe
Reviewed-on: https://review.monogon.dev/c/monogon/+/2795
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI